+++ pkg/drift_client/error.py
@@ -1,0 +1,5 @@
+"""DriftClient Errors"""
+
+
+class DriftClientError(Exception):
+    """A preliminary generic error"""